requirements.txt
connexion[swagger-ui]
setuptools
zappa>0.48.0
python-jose[cryptography]
six>=1.9
Flask>=0.10.1
boto3
flask_cors
